\name{ATTglasso-package}
\alias{ATTglasso-package}
\alias{ATTglasso}
\docType{package}
\title{
\packageTitle{ATTglasso}
}
\description{
\packageDescription{ATTglasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ATTglasso}
\packageIndices{ATTglasso}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ATTglasso}

Maintainer: \packageMaintainer{ATTglasso}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
}
\examples{
~~ simple examples of the most important functions ~~
}
