% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glasso_autotune}
\alias{glasso_autotune}
\title{Autotune Graphical Lasso}
\usage{
glasso_autotune(
  X,
  alpha = 0.1,
  penalize_diag = FALSE,
  thr = 1e-04,
  maxit = 10000L,
  verbose = TRUE,
  verbose_i = FALSE
)
}
\arguments{
\item{X}{Data matrix}

\item{alpha}{alpha value of F test}

\item{thr}{Threshold for convergence. Default value is 1e-4.}

\item{maxit}{Maximum number of iterations of outer loop. Default 10,000}
}
\value{
Estimated precision matrix
}
\description{
Autotune Graphical Lasso
}
