\name{large-package}
\alias{large-package}
\alias{large}
\docType{package}
\title{
\packageTitle{large}
}
\description{
\packageDescription{large}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{large}
\packageIndices{large}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{large}

Maintainer: \packageMaintainer{large}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
}
\examples{
~~ simple examples of the most important functions ~~
}
