% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{large}
\alias{large}
\title{Autotune Graphical Lasso}
\usage{
large(
  X,
  alpha = 0.02,
  penalize_diag = FALSE,
  thr = 0.05,
  maxit = 50L,
  verbose = TRUE,
  verbose_i = FALSE
)
}
\arguments{
\item{X}{Data matrix}

\item{alpha}{alpha value of F test. Default value is 0.02.}

\item{thr}{Threshold for convergence. Default value is 1e-4. Iterations stop when average absolute parameter change is less than thr * ave(abs(offdiag(s)))}

\item{maxit}{Maximum number of iterations of outer loop. Default 100.}
}
\value{
Estimated precision matrix
}
\description{
Autotune Graphical Lasso
}
