% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasso_autotune}
\alias{lasso_autotune}
\title{Get an initial set of putative variables for the GAM algorithm}
\usage{
lasso_autotune(
  X_X,
  X_Y,
  sigma2,
  n,
  s_22,
  y,
  Z,
  node,
  outer_iter,
  verbose,
  lambda0 = -1,
  alpha = 0.1
)
}
\arguments{
\item{X_X}{matrix of putative variables}

\item{X_Y}{the (possibly transformed) response}

\item{sigma2}{a threshold of (absolute) correlation above which a pair is considered highly correlated}
}
\value{
a list containing  variables to ignore because they are highly correlated with other, and SLR coefficients
}
\description{
Get an initial set of putative variables for the GAM algorithm
}
