% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glasso_autotune}
\alias{glasso_autotune}
\title{Graphical Lasso with Coordinate Descent and Autotuning}
\usage{
glasso_autotune(X, alpha = 0.1, thr = 1e-04, maxit = 10000L, verbose = TRUE)
}
\arguments{
\item{X}{matrix of putative variables}

\item{alpha}{the (possibly transformed) response}

\item{maxit}{a threshold of (absolute) correlation above which a pair is considered highly correlated}
}
\value{
a list containing  variables to ignore because they are highly correlated with other, and SLR coefficients
}
\description{
Graphical Lasso with Coordinate Descent and Autotuning
}
